ci: #1465 Phase A — SSOT CI スコープを src/**/*.svelte + site/**/*.html に拡張
- eslint.config.js: src/lib/**/*.svelte (stories 含む) に no-hardcoded-jp-text 適用
- check-hardcoded-strings.mjs: glob を src/**/*.svelte に拡張、HTML 検査を分離
- check-lp-ssot.mjs: 新設 — site/**/*.html 全対象 (法的文書含む) の SSOT 検査
- lp-ssot-baseline.json: 新設 — baseline 921 (既存 LP HTML 全件)
- hardcoded-strings-baseline.json: baseline 500 に更新 (stories 込み)
- ci.yml: check-lp-ssot.mjs を LP SSOT ジョブとして追加
