+++ .github/workflows/ci.yml
@@ -98,8 +98,11 @@ jobs:
       - name: No waitForTimeout in scripts/ check (#1208)
         run: node scripts/check-no-waitfortimeout.mjs
 
-      - name: Hardcoded JP text baseline check (#1452 Phase A)
+      - name: Hardcoded JP text baseline check (#1452 Phase A / #1465 Phase A)
         run: node scripts/check-hardcoded-strings.mjs
+
+      - name: LP SSOT HTML baseline check (#1465 Phase A)
+        run: node scripts/check-lp-ssot.mjs
 
       - name: No demo route duplication check (ADR-0039 / #1180)
         run: npm run check:no-demo-route-dup
