+++ scripts/hardcoded-strings-baseline.json
@@ -1,8 +1,6 @@
 {
-	"count": 0,
+	"count": 500,
 	"rule": "local/no-hardcoded-jp-text",
-	"scope": "src/routes/**/*.svelte",
-	"htmlCount": 865,
-	"htmlScope": "site/*.html",
-	"note": "Issue #1452 Phase C: all hardcoded JP strings migrated to labels.ts SSOT (0 violations). Rule escalated from warn to error. Issue #1464 Phase D: HTML check added with baseline of 865 (existing LP content). New HTML content must use data-lp-key attributes. Migrate existing content to reduce htmlCount to 0."
+	"scope": "src/**/*.svelte (stories included)",
+	"note": "Issue #1452 Phase C: src/routes 0 violations. Issue #1465 Phase A: scope expanded to src/**/*.svelte including stories — baseline 500. New code must not add violations. Migrate existing to reduce count to 0 (Phase B)."
 }
