+++ scripts/lp-ssot-baseline.json
@@ -1,0 +1,5 @@
+{
+	"count": 921,
+	"scope": "site/**/*.html",
+	"note": "Issue #1465 Phase A: LP SSOT HTML check baseline. Includes all site HTML (LP, legal docs). New HTML content must use data-lp-key attributes. Migrate existing content to reduce count to 0 (Phase C)."
+}
